backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
fpdf2
pydantic
python-dotenv
google-cloud-aiplatform
gunicorn
